# Improve performance of cell_length