--- a/rich/cells.py
+++ b/rich/cells.py
@@ -1,5 +1,5 @@
-from functools import lru_cache
 import re
+from functools import lru_cache
 from typing import Dict, List
 
 from ._cell_widths import CELL_WIDTHS
@@ -18,17 +18,13 @@ def cell_len(text: str, _cache: Dict[str, int] = LRUCache(1024 * 4)) -> int:
     Returns:
         int: Get the number of cells required to display text.
     """
+    cached_result = _cache.get(text, None)
+    if cached_result is not None:
+        return cached_result
 
-    if _is_single_cell_widths(text):
-        return len(text)
-    else:
-        cached_result = _cache.get(text, None)
-        if cached_result is not None:
-            return cached_result
-        _get_size = get_character_cell_size
-        total_size = sum(_get_size(character) for character in text)
-        if len(text) <= 64:
-            _cache[text] = total_size
+    _get_size = get_character_cell_size
+    total_size = sum(_get_size(character) for character in text)
+    _cache[text] = total_size
     return total_size
 
 
@@ -42,9 +38,6 @@ def get_character_cell_size(character: str) -> int:
     Returns:
         int: Number of cells (0, 1 or 2) occupied by that character.
     """
-    if _is_single_cell_widths(character):
-        return 1
-
     return _get_codepoint_cell_size(ord(character))