## Type of changes

- [ ] Bug fix
- [ ] New feature
- [ ] Documentation / docstrings
- [ ] Tests
- [x] Other - performance related change

## Checklist

- [x] I've run the latest [black](https://github.com/psf/black) with default args on new code.
- [x] I've updated CHANGELOG.md and CONTRIBUTORS.md where appropriate.
- [ ] I've added tests for new code.
- [x] I accept that @willmcgugan may be pedantic in the code review.

## Description

5-6% performance improvement on `python -m rich`

Removes the use of the regex for detecting codepoints in ranges we know to be single characters. It was consistently slower regardless of where cache lookups happened.

I've also removed the bounds on key length on the cell_len cache, which resulted in another immediately noticeable improvement.

`master`:
```
❯ hyperfine "python -m rich"
Benchmark 1: python -m rich
  Time (mean ± σ):     144.6 ms ±   1.5 ms    [User: 135.5 ms, System: 8.2 ms]
  Range (min … max):   143.5 ms … 150.6 ms    20 runs
```

this branch:
```
❯ hyperfine "python -m rich"
Benchmark 1: python -m rich
  Time (mean ± σ):     138.8 ms ±   1.0 ms    [User: 129.6 ms, System: 8.1 ms]
  Range (min … max):   137.3 ms … 142.0 ms    21 runs
```

Combining this change and the other PR which adjusted regex and caching - cumulative time spend inside `cell_len` during `python -m rich` has dropped from 0.02047 to 0.0056. See image below (top is old, bottom is new).

<img width="841" alt="image" src="https://user-images.githubusercontent.com/5740731/158200572-5496169a-1434-438d-ac89-24785280f22e.png">